BREAKING CHANGE: template state machine
